FEATURE: consolidates boards data/types into one file
