+++ libs/spirit-islander/util/src/lib/functions/create-model.ts
@@ -4,7 +4,7 @@ import { MAPS } from '../data/maps';
 import { SCENARIOS, ScenarioName } from '../data/scenarios';
 import { SPIRITS, SpiritName } from '../data/spirits';
 import type { AdversaryLevelId } from '../types/game/adversaries';
-import type { BalancedBoardName } from '../types/game/board';
+import type { BalancedBoardName } from '../data/boards';
 import type { ExpansionName, ExpansionOption } from '../types/game/expansions';
 import type { MapName } from '../types/game/maps';
 import { getOptionsByExpansion } from './get-options';
