--- libs/spirit-islander/util/src/lib/types/game/board.ts
@@ -1,15 +1,0 @@
-import type { ExpansionOption } from './expansions';
-
-export type BalancedBoardName = 'A' | 'B' | 'C' | 'D' | 'E' | 'F';
-
-export type ThematicBoardName =
-  | 'Northeast'
-  | 'East'
-  | 'Northwest'
-  | 'West'
-  | 'Southeast'
-  | 'Southwest';
-
-export interface Board extends ExpansionOption<BalancedBoardName> {
-  thematicName: ThematicBoardName;
-}
